# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Surebet Detector API",
    "description": "Real-time arbitrage and +EV detection across sportsbooks and prediction markets",
    "version": "2.0.0",
    "contact": {
      "name": "Surebet Team",
      "email": "api@surebet-detector.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.surebet-detector.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": ["Events"],
        "summary": "List all events",
        "description": "Returns all active sports events with aggregated odds",
        "operationId": "listEvents",
        "parameters": [
          {
            "name": "sport",
            "in": "query",
            "description": "Filter by sport",
            "schema": {
              "type": "string",
              "enum": ["soccer", "tennis", "basketball", "baseball", "hockey", "esports"]
            }
          },
          {
            "name": "league",
            "in": "query",
            "description": "Filter by league",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of events to return",
            "schema": {
              "type": "integer",
              "default": 100,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{eventId}": {
      "get": {
        "tags": ["Events"],
        "summary": "Get event details",
        "description": "Returns detailed odds information for a specific event",
        "operationId": "getEvent",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "description": "Unique event identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregatedOdds"
                }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventsResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AggregatedOdds"
            }
          }
        }
      },
      "AggregatedOdds": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string"
          },
          "sport": {
            "type": "string"
          },
          "league": {
            "type": "string"
          },
          "homeTeam": {
            "type": "string"
          },
          "awayTeam": {
            "type": "string"
          },
          "startTime": {
            "type": "integer",
            "description": "Unix timestamp"
          },
          "markets": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": {
                "type": "object",
                "additionalProperties": {
                  "$ref": "#/components/schemas/OddsEntry"
                }
              }
            }
          },
          "bestOdds": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": {
                "$ref": "#/components/schemas/BestOddsEntry"
              }
            }
          },
          "lastUpdated": {
            "type": "integer",
            "description": "Unix timestamp"
          }
        }
      },
      "OddsEntry": {
        "type": "object",
        "properties": {
          "odds": {
            "type": "number"
          },
          "timestamp": {
            "type": "integer"
          },
          "volume": {
            "type": "number",
            "nullable": true
          }
        }
      },
      "BestOddsEntry": {
        "type": "object",
        "properties": {
          "bookmaker": {
            "type": "string"
          },
          "odds": {
            "type": "number"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}